refactor: use the new shared ConnectionInformation component
